fix(yip): student vote/bill/questions pages could not read httpOnly session
The yip_session cookie is httpOnly (auth.ts), but /yip/me/vote, /yip/me/bill
and /yip/me/questions parsed it client-side via document.cookie — which never
sees httpOnly cookies. Every logged-in student got 'Session not found. Please
rejoin the event.' on all three pages, blocking party-leader/speaker voting,
bill drafting and Question Hour submissions from student phones.

Fix: split each page into a server page.tsx that reads the cookie via
next/headers (redirect to /yip/join when absent) + a *-client.tsx receiving
the session as a prop — the same pattern app/yip/me/motion already uses.
